feat: generating classes days

+++ src/ClassManager.Domain/Contexts/TimeTable/Handlers/UpdateTimeTableHandler.cs
@@ -51,11 +51,10 @@ public class UpdateTimetableHandler :
       return new CommandResult(false, "ERR_ADMIN_ROLE_NOT_FOUND", null, null, 403);
     }
 
-    var timeTable = await _timeTableRepository.FindAsync(x => x.Id == timeTableId && x.TenantId == tenantId, [x => x.Classes, x => x.SchedulesDays]);
+    var timeTableExists = await _timeTableRepository.IdExistsAsync(timeTableId, tenantId, new CancellationToken());
 
-    if (timeTable is null)
+    if (!timeTableExists)
     {
-      return new CommandResult(false, "ERR_TIME_TABLE_NOT_FOUND", null, null, 404);
     }
 
     List<ScheduleDay> schedulesDaysEntities = [];
@@ -70,13 +69,21 @@ public class UpdateTimetableHandler :
 
     await _scheduleDayRepository.CreateRangeAsync(schedulesDaysEntities, new CancellationToken());
 
-    var classesIds = timeTable.Classes.Select(c => c.Id).ToList();
 
     var lastDayOfMonth = DateTime.Now.GetLastDayOfMonth().AddHours(23).AddMinutes(59).AddSeconds(59);
 
+    var timeTableReloaded = await _timeTableRepository.FindAsync(x => x.Id == timeTableId && x.TenantId == tenantId, [x => x.Classes, x => x.SchedulesDays]);
+
+    if (timeTableReloaded is null)
+    {
+      return new CommandResult(false, "ERR_TIME_TABLE_NOT_FOUND", null, null, 404);
+    }
+
+    var classesIds = timeTableReloaded.Classes.Select(c => c.Id).ToList();
+
     await _classDayRepository.DeleteAllAfterAndBeforeDate(classesIds, DateTime.Now, lastDayOfMonth, new CancellationToken());
 
-    await _generateClassesDaysHelper.Handle([timeTable], DateTime.Now.Year, DateTime.Now.Month, DateTime.Now.Day);
+    await _generateClassesDaysHelper.Handle([timeTableReloaded], DateTime.Now.Year, DateTime.Now.Month, DateTime.Now.Day);
 
     return new CommandResult(true, "TIME_TABLE_UPDATED", "", null, 200);
   }
